feat: rename LicensesPage and LicensesView components and update import paths
